Adopt `MontgomeryXpoint` naming scheme
